Reducing any possible typos and potential code issues

+++ dlls/bot_commands.cpp
@@ -1878,10 +1878,10 @@ eBotCvarState BotFunc_AddBot(CClient* pClient, const char* arg1, const char* arg
 
 		const char* teamlist = CVAR_GET_STRING("mp_teamlist");
 
-		int i = 0;
-		int j = 0;
+		unsigned int i = 0;
+		unsigned int j = 0;
 
-		const int len = std::strlen(teamlist);
+		const size_t len = std::strlen(teamlist);
 
 		while (i < len)
 		{
